transform ConnectionStateMixin in HarmonyEntity base class
