+++ homeassistant/components/harmony/switch.py
@@ -5,9 +5,9 @@ from homeassistant.components.switch import SwitchEntity
 from homeassistant.const import CONF_NAME
 from homeassistant.core import callback
 
-from .connection_state import ConnectionStateMixin
 from .const import DOMAIN, HARMONY_DATA
 from .data import HarmonyData
+from .entity import HarmonyEntity
 from .subscriber import HarmonyCallback
 
 _LOGGER = logging.getLogger(__name__)
@@ -27,18 +27,16 @@ async def async_setup_entry(hass, entry, async_add_entities):
     async_add_entities(switches, True)
 
 
-class HarmonyActivitySwitch(ConnectionStateMixin, SwitchEntity):
+class HarmonyActivitySwitch(HarmonyEntity, SwitchEntity):
     """Switch representation of a Harmony activity."""
 
     def __init__(self, name: str, activity: dict, data: HarmonyData) -> None:
         """Initialize HarmonyActivitySwitch class."""
-        super().__init__()
+        super().__init__(data=data)
         self._name = name
         self._activity_name = activity["label"]
         self._activity_id = activity["id"]
-        self._data = data
         self._attr_entity_registry_enabled_default = False
-        self._attr_should_poll = False
         self._attr_unique_id = f"activity_{self._activity_id}"
         self._attr_name = name
         self._attr_device_info = self._data.device_info(DOMAIN)
@@ -48,11 +46,6 @@ class HarmonyActivitySwitch(ConnectionStateMixin, SwitchEntity):
         """Return if the current activity is the one for this switch."""
         _, activity_name = self._data.current_activity
         return activity_name == self._activity_name
-
-    @property
-    def available(self):
-        """Return True if we're connected to the Hub, otherwise False."""
-        return self._data.available
 
     async def async_turn_on(self, **kwargs):
         """Start this activity."""
